refactor(ui): Move toast modifiers to a separate ViewExtension file
