+++ PlayCover/Views/MainView.swift
@@ -181,23 +181,6 @@ struct MainView: View {
     }
 }
 
-extension View {
-    @ViewBuilder
-    func toastOverlay<Content: View>(
-        @ViewBuilder content: () -> Content
-    ) -> some View {
-        if #available(macOS 26.0, *) {
-            self.safeAreaBar(edge: .bottom) {
-                content()
-            }
-        } else {
-            self.overlay {
-                content()
-            }
-        }
-    }
-}
-
 struct SplitViewAccessor: NSViewRepresentable {
     @Binding var sideCollapsed: Bool
